feat: Add device option for translation
The code changes add a new `--device` option to the `translation` command in the `cli.py` file. This option allows users to specify the device to use for inference when translating text. The default value is set to "auto", which sets the device map to "auto" in the `SimpleGenerator` constructor.